+++ src/Modules/Admin/Widgets/Forms/AssetActiveForm.php
@@ -9,6 +9,7 @@ use Hirtz\Cms\Modules\Admin\Widgets\Forms\Traits\ActiveFormFieldsTrait;
 use Hirtz\Media\Modules\Admin\Widgets\Forms\Traits\AssetFieldsTrait;
 use Hirtz\Skeleton\Widgets\Forms\ActiveForm;
 use Hirtz\Skeleton\Widgets\Forms\Fields\InputField;
+use Hirtz\Skeleton\Widgets\Forms\Traits\CustomAttributeFieldsTrait;
 use Override;
 use Stringable;
 
@@ -19,6 +20,7 @@ class AssetActiveForm extends ActiveForm
 {
     use ActiveFormFieldsTrait;
     use AssetFieldsTrait;
+    use CustomAttributeFieldsTrait;
 
     #[Override]
     protected function configure(): void
@@ -35,6 +37,7 @@ class AssetActiveForm extends ActiveForm
                 $this->getAltTextField(),
                 $this->getLinkField(),
                 $this->getEmbedUrlField(),
+                ...$this->getCustomAttributeFields(),
             ],
         ];